Naive add and remove index added
You can now naively add and remove indexes. Added a PostgreSQL extension to our DSL that lets you add an index on a table and a set of it's columns.  You can also specify binary flags and any non-binary flags with a default (right now this is just ASC / DESC).

If an index on these columns already exists, you'll get an index name like "tablename_column1_column2_idx1" in postgres. There isn't any checking for this, so for now using this you'd be careful not to create two indexes on the same table/columns, you'll have to manually delete any above the first one and add-index :down will throw unexpected errors.